+++ ave/__init__.py
@@ -1,5 +1,5 @@
 from .ave import AVE  # noqa: F401
 from .game_loader import load_game_from_library  # noqa: F401
 from .game_loader import load_game_from_file  # noqa: F401
-from .game import Character, Game, Item  # noqa: F401
+from .game import Character, Game  # noqa: F401
 from .screen import Screen, DummyScreen  # noqa: F401
