+++ ave/game_loader.py
@@ -1,10 +1,11 @@
 import re
 import json
 import urllib.request
-from .game import Game, Room, Item, Number
+from .game import Game, Room
 from .game import (TextWithRequirements, OptionWithRequirements,
                    NameWithRequirements)
 from .escaping import escape, unescape, clean, between
+from .items import Item, Number
 from . import requirements as rq
 from . import item_giver as ig
 from . import numbers as no
